DiLoCo: move the upload wire-cast into the backend (step 3 of #154) (#157)
Step 3 of the #154 role decomposition. Moves the upload wire-precision cast out
of ParamView/FragmentManager into HttpStarBackend, so each backend owns its own
wire representation (shared-memory = no cast; collective = packed fp8/fp4, per
#153) without touching the model abstraction.

Behavior-preserving: the wire bytes (dtype + stochastic rounding) are identical;
only where the worker-side cast runs changes. The download cast was already
server-side and is untouched.

- wire_cast.py (new): cast_for_upload, moved verbatim from
  param_view._cast_for_upload (SR semantics preserved).
- param_view.py / fragments.py: compute_pseudograds returns the raw difference in
  the live model dtype; drop the upload_dtype/upload_sr params.
- sync_backend.py: HttpStarBackend takes upload_dtype/upload_sr and casts the raw
  pseudo-grads in synchronize / synchronize_fragment; SyncResult gains
  sent_bytes/recv_bytes (the backend owns the wire repr, so it reports the size).
- worker.py: build the default backend with the negotiated upload config; pass
  raw pseudo-grads; take the wire-byte metrics from the SyncResult.

Verified: full tests/unit/ml/diloco/ suite; live smoke confirms
diloco/last_send_mb still reports the bf16 wire size (68.8 MB, not fp32 137 MB).

+++ tests/unit/ml/diloco/test_streaming.py
@@ -151,10 +151,8 @@ class TestFragmentManager(unittest.TestCase):
             for p in model.parameters():
                 p.add_(torch.ones_like(p) * 0.5)
 
-        # Compute pseudo-grads for fragment 0 only
-        pseudograds = fm.compute_fragment_pseudogradients(
-            0, global_params, model, upload_dtype="fp32"
-        )
+        # Compute pseudo-grads for fragment 0 only (raw; the backend casts)
+        pseudograds = fm.compute_fragment_pseudogradients(0, global_params, model)
 
         # Only fragment 0's params should be present
         frag0_names = set(fm.fragments[0])
